webapp/requirements.txt
Flask==0.10.1
cqlengine==0.21.0
kafka-python==0.9.2
